Move hashing functionality to its own file
